rename module table -> table_pane
